Add ChartChrome component
Adds a wrapper component which renders axes and the legend. Provides
useful helper functions, such as series toggling and fixation, through
the ChartContext.

Child components can then receive the data from the context and decide
how to render it.

Currently, this component is limited by the ScaleBand x-axis, and
support for continuous scales is required to unlock proper support for
area and line charts.